+++ README.md
@@ -1,4 +1,4 @@
-# Practical-6
+# Practical-5
 ## User List App
 #### 1. CRA
 
@@ -37,13 +37,7 @@ $ npm install @reduxjs/toolkit
 ##### Open http://localhost: .... to view it in your browser.
 
 #### App Preview:
-![Screenshot from 2022-03-29 10-13-08](https://user-images.githubusercontent.com/97106864/160534896-be85dd57-aec1-4a0b-8240-728990c3f63a.png)
-![Screenshot from 2022-03-29 10-05-11](https://user-images.githubusercontent.com/97106864/160534412-0c31c3c1-009e-4b68-9955-f731cb237e33.png)
-![Screenshot from 2022-03-29 10-06-02](https://user-images.githubusercontent.com/97106864/160534449-ac41ac50-602d-404c-9737-aba8cdc90bd6.png)
-![Screenshot from 2022-03-29 10-07-12](https://user-images.githubusercontent.com/97106864/160534466-74ad0e27-e9e9-4ee2-b615-27b3478d0ad8.png)
-
-
-
-
+![Screenshot from 2022-03-24 16-28-27](https://user-images.githubusercontent.com/97106864/159902174-4be78868-a47e-4743-9de9-eaebeb48f08c.png)
+![Screenshot from 2022-03-24 16-29-01](https://user-images.githubusercontent.com/97106864/159902243-c9944a73-248c-4348-917a-618d7ee2d6a3.png)
